% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_unifed_glm.R
\name{summary_unifed_glm}
\alias{summary_unifed_glm}
\title{Summarizing Generalized Linear Model Fits}
\usage{
summary_unifed_glm(object, ...)
}
\arguments{
\item{object}{an object of class "glm".}

\item{...}{Other arguments for \code{stats::summary.glm}.

This wrapper function was created in order to automatically set to
1 the dispersion parameter of a fitted unifed GLM. When the package
is loaded the summary method of the glm class is rewritten using
this function.}
}
\description{
Wrapper function for summary.glm.
}
