% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifed.mle.R
\name{unifed.mle}
\alias{unifed.mle}
\title{Maximum Likelihood Estimate for the unifed distribution}
\usage{
unifed.mle(x)
}
\arguments{
\item{x}{A numeric vector with values in the interval [0,1].

Computes the maximum likelihood estimator of the canonical
parameter of the unifed distribution. It is assumed that the
elements of \code{x} come from independent and identically
distributed unifed random variables.}
}
\description{
Maximum Likelihood Estimate for the unifed distribution
}
\examples{
a.unifed.sample <- runifed(1000,10)
theta.mle <- unifed.mle(a.unifed.sample)

}
