% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifed.stan.path.R
\name{unifed.stan.path}
\alias{unifed.stan.path}
\alias{unifed.stan.folder}
\title{Unifed Stan function paths}
\usage{
unifed.stan.path()

unifed.stan.folder()
}
\value{
The full path to the unifed.stan file provided by the
    package.

\code{unifed.stan.folder} returns a string containing the
    path to the folder containing the unifed.stan file. This can be
    used as the \code{isystem} parameter in stan functions.
}
\description{
The unifed.stan provided by the file contains functions for using
the unifed distribution in stan. The file can be included (with
#include) insided the functions block of a stan program or its
contents can be copied and pasted.
}
