% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifed.varf.R
\name{unifed.varf}
\alias{unifed.varf}
\title{Variance function of the unifed distribution}
\usage{
unifed.varf(mu)
}
\arguments{
\item{mu}{A vector with numbers between 0 and 1.}
}
\value{
It returns
    \code{unifed.kappa.double.prime(unifed.kappa.prime.inverse(mu))}.
}
\description{
Variance function of the unifed distribution
}
